+++ src/app/modules/shared/components/file-upload-button/file-upload-button.component.ts
@@ -16,6 +16,9 @@ export class FileUploadButtonComponent {
     @Input()
     public label: string | null = null;
 
+    @Input()
+    public help: string | null = null;
+
     @Output()
     public upload = new EventEmitter<string>();